backend/core/models/_base.py
from django.contrib.auth.models import (
    AbstractBaseUser,
    BaseUserManager,
    PermissionsMixin,
)
from django.db import models
import uuid
from phonenumber_field.modelfields import PhoneNumberField
